move to hook useAutoTheme
